refactor: 가독성을 위해 메인 함수로직 분할해서 새로운 함수 만들어서 실행(playGame)
playGame()함수를 새로 만들어서 파일 분리
